pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "demo-agents"
version = "0.2.8"
requires-python = ">=3.9"

dependencies = [
    # 核心依赖
    "openai>=1.0.0,<2.0.0",
    "python-dotenv>=0.19.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi",
    "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
# 搜索功能依赖
search = [
    "tavily-python>=0.7.12"
]

[tool.setuptools.packages.find]
where = ["src"]

